packages/mathapi/requirements.txt
Flask>=2.2.2
Werkzeug>=2.2.2
requests
python-dotenv
pymongo
dnspython  # Required if you're using MongoDB Atlas or a connection string with DNS
openai>=1.58.1